requirements.txt
Django~=6.0.2
djangorestframework~=3.16.0
PyJWT~=2.10.1
python-dotenv~=1.2.1
psycopg2-binary
whitenoise~=6.9.0